requirements.txt
# Основные зависимости
python-telegram-bot==20.7
aiogram==3.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic==2.5.0
pydantic[email]>=2.4.0
pydantic-settings>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.12.0
psycopg2-binary==2.9.9

# AI и обработка данных
openai>=1.3.0
whisper>=1.1.10
easyocr>=1.7.0
numpy>=1.24.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Векторный поиск и эмбеддинги
sentence-transformers>=2.2.0
pgvector>=0.2.0
faiss-cpu>=1.7.0

# Аутентификация и безопасность
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=41.0.0
bcrypt==4.1.2

# Кэширование и производительность
redis>=5.0.0
redis[hiredis]==5.0.1
celery[redis]>=5.3.0
aioredis==2.0.1

# Утилиты
python-dotenv>=1.0.0
aiofiles>=23.0.0
aiohttp>=3.9.0
Pillow==10.1.0
requests==2.31.0
httpx>=0.25.0

# Логирование и мониторинг
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client>=0.19.0
loguru>=0.7.0

# Тестирование
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0

# Разработка
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Календарь и планирование
icalendar==5.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
msal==1.24.1

# Уведомления и интеграции
jinja2==3.1.2

# Аналитика и отчеты
reportlab==4.0.7
xlsxwriter==3.1.9
openpyxl==3.1.2

# Оптимизация и мониторинг базы данных
sqlalchemy-utils==0.41.1

# Безопасность и валидация
email-validator>=2.1.0
phonenumbers>=8.13.0

# Метрики и профилирование
memory-profiler==0.61.0
psutil==5.9.6

# CI/CD и развертывание
gunicorn==21.2.0
supervisor==4.2.5

# Дата и время
pytz>=2023.3

# Парсинг
beautifulsoup4>=4.12.0
lxml>=4.9.0 